Added Show API to Keystone
Most commands start with a GET /v2.0 request on port 5001. Thus i added it to avoid 404 errors.
neutron *-show commands don't add a <network_id> to the request but rather add '?fields=...' with the name or id of a Network, Subnet or Port. At the moment check the requests for such parameters and call the function if necessary.
neutron subnet-list will now display the subnets with the correct allocation_pools.