DeleteNetwork works (but not the neutron client)
Added Create functions for Net, Subnet and Port.
Some style updates.
Creating a Net works - rest was not completly tested.
If you delete a subnet and not the net - know that the neutron client will always ask for subnets when you want to list the nets! Thus it will throw errers if you only delete a subnet and dont add a new one!