Finally the neutron client can create, delete and show - networks etc.
The answer of the keystone_api will now contain the version (v2.0).
Entry points for the neutron client where wrong - now they should work.
Individual net, subnet and port classes can be requested with the neutron client.
The Delete functions should work now and the delete function for ports was added.
A link will still NOT be created when the net, subnet and port are created - we have to figure out which server should be used and how we specify the two connection points of this link!