Interface names should no longer change after update
Fixed missing ']' in chain api.
Removed interface name checking - because we create the interfaces ourselves. Thus we are responsible for all interface names and to update old stacks i need to create the names after we have checked the stack.
Heat parser now also takes an argument, if it is an stack update or a new stack - thus i know when to create everything and when only create important stuff.
Each port will now create its own uuid.
Used interface names 'intf_names' will now use a list of port.id and boolean to determin which port has which interface name and if this interface is reused while updating the stack or if it is a new one. Thus we know if it is no longer used and if we can delete it.
'update_intf_name()' is no longer needed because 'create_intf_name()' will check if it can reuse older interface names.